feat: expand DataTag to uint32

+++ cli/serverless/wasm/runtime.go
@@ -20,10 +20,10 @@ type Runtime interface {
 	Init(wasmFile string) error
 
 	// GetObserveDataTags returns observed datatags of the wasm sfn
-	GetObserveDataTags() []byte
+	GetObserveDataTags() []uint32
 
 	// RunHandler runs the wasm application (request -> response mode)
-	RunHandler(data []byte) (byte, []byte, error)
+	RunHandler(data []byte) (uint32, []byte, error)
 
 	// Close releases all the resources related to the runtime
 	Close() error
