+++ example/4-cascading-zipper/sfn/sfn_echo.go
@@ -29,7 +29,7 @@ func main() {
 	select {}
 }
 
-func handler(data []byte) (byte, []byte) {
+func handler(data []byte) (uint32, []byte) {
 	val := string(data)
 	log.Printf(">> [sfn] got tag=0x33, data=%s", val)
 	return 0x0, nil
