+++ example/7-wasm/sfn/zig/src/main.zig
@@ -1,8 +1,8 @@
 const std = @import("std");
 
-extern fn yomo_observe_datatag(tag: u8) void;
+extern fn yomo_observe_datatag(tag: u32) void;
 extern fn yomo_load_input(pointer: *const u8) void;
-extern fn yomo_dump_output(tag: u8, pointer: *const u8, length: usize) void;
+extern fn yomo_dump_output(tag: u32, pointer: *const u8, length: usize) void;
 
 pub fn main() anyerror!void {
     std.log.info("yomo wasm sfn on zig", .{});
