+++ example/9-cli/sfn-db/app.go
@@ -13,7 +13,7 @@ type NoiseData struct {
 }
 
 // Handler will handle the raw data.
-func Handler(data []byte) (byte, []byte) {
+func Handler(data []byte) (uint32, []byte) {
 	// var noise float32
 	var noise NoiseData
 	err := json.Unmarshal(data, &noise)
@@ -27,6 +27,6 @@ func Handler(data []byte) (byte, []byte) {
 	return 0x0, nil
 }
 
-func DataTags() []byte {
-	return []byte{0x34}
+func DataTags() []uint32 {
+	return []uint32{0x34}
 }
